requirements.txt
# Core ML 
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Experiment tracking & model registry 
mlflow==2.16.2

# Model serving
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.3

# Monitoring & drift detection 

evidently==0.7.21

# Dashboard UI
streamlit==1.29.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
requests
httpx
joblib==1.3.2
PyYAML==6.0.1
rich==13.7.0
click==8.1.7
schedule==1.2.1

# Testing 
pytest==7.4.4